gltf_loader_interfaces.cpp: correct 1 cast issue

+++ src/Loaders/src/loading/glTF/2.0/gltf_loader_interfaces.cpp
@@ -113,7 +113,7 @@ IGLTF2::IAccessorSparseValues::Parse(const json& parsedAccessorSparseValues)
   if (json_util::has_valid_key_value(parsedAccessorSparseValues,
                                      "byteOffset")) {
     accessorSparseValues.byteOffset
-      = json_util::get_number<size_t>(parsedAccessorSparseValues, "byteOffset");
+      = json_util::get_number<int>(parsedAccessorSparseValues, "byteOffset");
   }
 
   return accessorSparseValues;
